+++ src/containers/Appointments.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import User from './User';
+import ManageAppointments from '../pages/ManageAppointments';
+
+export default function Appointments() {
+  return (
+    <User>
+      <ManageAppointments />
+    </User>
+  );
+}
